// ParkingSystem.UI/src/app/Models/parkingDetail.model.ts
export interface parkingDetail {
  zoneId: number,
  zoneTitle: string;
  totalSlots: number;
  filledSlots: number;
  availableSlots: number;
  floorId: number;
  allSlots : Array<number>;
  slot1:  number;
  slot2:  number;
  slot3:  number;
  slot4:  number;
  slot5:  number;
  slot6:  number;
  slot7:  number;
  slot8:  number;
  slot9:  number;
  slot10: number;
  slot11: number;
  slot12: number;
  slot13: number;
  slot14: number;
  slot15: number;
  slot16: number;
  slot17: number;
  slot18: number;
  slot19: number;
  slot20: number;
}